pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "multimodal-learning-tumor-classification"
version = "0.1.0"
description = "Multimodal breast cancer tumor grading from DCE-MRI and clinical features"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "pydicom>=2.3.0",
    "scikit-learn>=1.2.0",
    "pandas>=1.5.0",
    "openpyxl>=3.0.0",
    "matplotlib>=3.6.0",
    "pillow>=9.0.0",
    "numpy>=1.23.0",
]

[project.scripts]
tumor-classify = "multimodal_tumor_classification.__main__:main"

[tool.setuptools.packages.find]
include = ["multimodal_tumor_classification*"]
